requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
datasets==2.14.6
huggingface_hub==0.19.4
pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
wandb==0.16.0
